Fix issues and updated the files
